Develop

See merge request bity/dev/icp/gold-dao!601

+++ backend/canisters/gldt_swap/archive/src/state.rs
@@ -54,7 +54,7 @@ pub struct CanisterInfo {
     pub version: BuildVersion,
     pub commit_hash: String,
     pub memory_used: MemorySize,
-    pub cycles_balance_in_tc: f64,
+    pub cycles_balance_in_tc: u128,
 }
 
 #[derive(Serialize, Deserialize)]
