+++ backend/libraries/types/src/lib.rs
@@ -20,7 +20,7 @@ pub struct Empty {}
 
 pub type CanisterId = Principal;
 pub type CanisterWasm = Vec<u8>;
-pub type Cycles = u64;
+pub type Cycles = u128;
 pub type Hash = [u8; 32];
 pub type Maturity = u64;
 pub type Second = u64;
